fix: consolidate resize and mutation observers into a unified hook for improved carousel init
